Initial attempt at issue #5
This adds a `--offline` command-line argument that uses the local filesystem to generate a manifest rather than querying a registry service. Note that in this "offline" mode, we cannot read the bundle XML file from stdin; we need the filesystem location so we can locate the rest of the data files to generate hashes and the manifest. Using `--offline` prevents the use of `--url`. The resulting SIP manifest gets written to `stdout` as before.

This commit also includes a complete mission bundle under `docs/samples/mission_bundle` for testing; since the `LADEE_Bundle_1101.xml` file is part of that directory tree, we get rid of the redundant one in `docs/samples`.

See issue #5 for a brief overview of the algorithm.