Attempt at #9
This switches things up to use lxml instead of Python's built-in ElementTree. The reason: none of the built-in XML tools let you put in an XML Processing Instruction at the start of the XML doc. (ElementTree does allow PIs, but only within the <root-element>, which is kind of silly.) Otherwise we change things up to require an input file (goodbye `stdin`) and write two output files (the manifest file and its accompanying label). The label gets populated with all sorts of XML goodness.