Resolve #77 and #78: use a temp dir instead of an unlinked temp file; fix arg to calling process on aipgen
Previously, we were using `NamedTemporaryFile` to make a temporary file for `sqlite3`. `sqlite3` would close and open this file several times during processing. On Unix, this is fine. On Windows NT or later, the very first "close" on the file would render it unusable by future opens. So instead, we make a temporary _directory_ instead of a file and use a known filename inside of that (#77).

Also, we noticed that there was a redundant `args.` being referenced in the call to `process` in `aip.py`. This fixes that (#78).