Fix secret detection against baseline
The problem wasn't that a new secret suddenly appeared. Instead, it was that the `.pre-commit-config.yaml` wasn't included in the baseline configuration which resulted in a diff when the GitHub Action ran. The GitHub Action from SLIM is a bit primitive: it treats _any diff_ as a new secret. In this case, the diff was a difference in config, not in secrets.

This also properly escapes the `.` in the regex for excluded files named `.pre-commit-config.yaml`.